Bus Map (#129)
* NEW BRANCH

added everything from original busMap branch but without the flamboiency.

* Got everything in the new branch updated and working theoretically

need to use position attribute to accurately display bus positions and work on bus lots.

* Make admin page for the bus map so they can input bus numbers for the lots

need to implement the actual saving the data part and color change for pinned buses.

* Work on how to update the lists that contain lot row info

continue working on that in router and client side and what not.

* Work on pin color highlights for bus map

will change the color of any pinned buses displayed on the map.

* Continue implementing the color change and lots

* Finished the highlight feature

remove test button for it and have it auto highlight.

* Highlight system stuff

* Bus map stuff

* Finished the glowing effect for the bus map for pinned buses.

* Create separate pages for the bus lots to simplify because otherwise page refreshes get real wack.

* Clean up how the highlights work as well as the update files for the bus map

need to implement the lots.

* Messed around with some stuff for the lots

* Look at updateBusList for reference on how to do the lots.

* Add popup for how the bus map works

finish styling everything, need to implement lots.

* Finished up with the info bar.

---------

Co-authored-by: myulbarsov <myulbarsov@stu.naperville203.org>
Co-authored-by: myulbarsov <156241499+myulbarsov@users.noreply.github.com>